fix: #wz-31189 review epic6 chunk 1 (CRITICAL + HIGH)
Adversarial review of the user-level-overlays branch (chunks 1a backend
core + 1b cross-cutting): 1 CRITICAL + 16 HIGH findings introduced by
this branch are addressed; pre-existing patterns are deferred.

CRITICAL
- UserGroupController endpoints were reachable by any authenticated user
  (IDOR cross-namespace + open enumeration). Add UserGroupGuard that
  resolves namespaceExternalId → namespace and gates with READ/WRITE
  permission; @PreAuthorize + HideOnAccessDenied for 404-not-403.

Backend core (chunk 1a)
- P1+P14+P17 pagination: Long arithmetic + drop dead branch in 3 user
  controllers (UserIntegrationConfig + UserAiModel + UserAiProvider).
- P3 IntegrationConfigServiceImpl: drop chained throwable from
  logger.warn to stop echoing Neo4j property values (NFR-SEC-4).
- P5 AiModel.findActiveByTriple: ORDER BY alias DESC, id ASC for
  deterministic LIMIT 1 across alias/apiName collisions.
- P6 IntegrationConfigMergeStrategy: explicit JSON null in override
  inherits base instead of wiping it (silent credential blanking on
  round-trip masked DTOs). +4 unit tests + KDoc IG-2/IG-3.
- P8 UserAiModelController: defensive null check on body.aiProviderId.
- P9 IntegrationConfigSchemaInitializer: assertNoBlankNames pre-flight
  before tripleKey backfill (Cypher + with NULL produces NULL, would
  trip the downstream constraint with a vague error).
- P11 RunReconciliationCache.CacheKey now includes (namespaceId, userId)
  as defence-in-depth in case the per-run lifecycle ever breaks. +2
  unit tests.
- P16 IntegrationConfigSchemaInitializer.assertNoTripleKeyDuplicates:
  pre-flight duplicate detection with sample keys before CREATE
  CONSTRAINT, replacing the opaque Neo4j error.

Cross-cutting (chunk 1b)
- H-1 InMemoryUserGroupRepository: real implementation via injected
  NamespaceRepository instead of returning emptyList() (POST in-memory
  no longer 500s on the now-removed re-fetch path).
- H-2 UserGroupServiceImpl.createFromRequest: return the freshly-created
  entity directly; remove racy re-fetch + .first { name }.
- H-3 NamespaceController: externalId immutable on PUT — prevents
  federation hijack/spoof by a namespace ADMIN. Settable only at create
  by SUPER_ADMIN.
- H-4 NamespaceServiceImpl.delete: cascade soft-delete UserGroup via
  injected UserGroupRepository (avoids cycle with UserGroupServiceImpl).
- H-5 Neo4jUserGroupRepository: tombstone (name, namespaceId) at
  soft-delete to free the unique constraint for re-creation. Same
  pattern as IntegrationConfig.tripleKey.
- H-6 UserGroupCreateRequest: drop userIds/agentIds (silently-ignored
  fields). Membership tracked as a future epic.
- H-7 Neo4jSchemaInitializer: ApplicationReadyEvent → ApplicationRunner
  with explicit fail-fast — a failure now aborts boot instead of
  reporting ready with constraints absent.
- H-8 fail-closed discrimination: ToolRegistryService.resolveToolsForRun
  now rethrows ConfigNotFoundException only for names enumerated from
  findByNamespaceShared (NFR-REL-1); names that come exclusively from
  user overrides remain silently skipped (FR30 dormant). KDoc on
  AgentServiceImpl.applyOverlaysToModel documents the symmetric
  posture (always fail-closed, baseModel exists by construction).

Tests
- IntegrationConfigMergeStrategySpec: +4 NullNode scenarios.
- RunReconciliationCacheSpec: +2 (namespaceId, userId) scoping scenarios.
- ToolRegistryServiceSpec: AC2 split into fail-closed shared vs dormant
  user override.
- UserGroupControllerIntegrationSpec: refactored with @MockkBean on
  UserService + PermissionService, auth-gated cases (404 without
  permission, 200 with).
- NamespaceServiceImplSpec: DI updated for the new UserGroupRepository.

+++ agentos/agentos-service/src/main/kotlin/io/whozoss/agentos/aiProvider/UserAiProviderController.kt
@@ -109,7 +109,7 @@ class UserAiProviderController(
             page = safePage,
             size = safeSize,
             totalElements = total.toLong(),
-            totalPages = (total + safeSize - 1) / safeSize,
+            totalPages = ((total.toLong() + safeSize - 1) / safeSize).toInt(),
         )
     }
